Move review app middleware into `common` directory
